fix(npm): support post install scripts being disabled
