+++ deployment/npm/bin.js
@@ -8,25 +8,38 @@ const fs = require("fs");
 
 const exePath = path.join(__dirname, os.platform() === "win32" ? "dprint.exe" : "dprint");
 
-try {
-  const result = child_process.spawnSync(
-    exePath,
-    process.argv.slice(2),
-    { stdio: "inherit" },
-  );
+if (!fs.existsSync(exePath)) {
+  require("./install_api").runInstall().then(() => {
+    runDprintExe();
+  }).catch(err => {
+    console.error(err);
+    process.exit(child_process.exitCode || 1);
+  });
+} else {
+  runDprintExe();
+}
 
-  if (result.status !== 0) {
+function runDprintExe() {
+  try {
+    const result = child_process.spawnSync(
+      exePath,
+      process.argv.slice(2),
+      { stdio: "inherit" },
+    );
+    if (result.error) {
+      throw result.error;
+    }
+
     throwIfNoExePath();
-  }
 
-  process.exitCode = result.status;
-} catch (err) {
-  throwIfNoExePath();
-  throw err;
+    process.exitCode = result.status;
+  } catch (err) {
+    throw err;
+  }
 }
 
 function throwIfNoExePath() {
   if (!fs.existsSync(exePath)) {
-    throw new Error("Could not find exe at path '" + exePath + "'. Please ensure the dprint 'postinstall' script runs on install.");
+    throw new Error("Could not find exe at path '" + exePath + "'. Maybe try running the dprint again.");
   }
 }
